+++ client/src/components/room/follower-select-room-modal.tsx
@@ -1,19 +1,17 @@
 /* eslint-disable */
-import React, { useEffect, useState, useRef } from 'react';
+import React, { useRef } from 'react';
 import styled from 'styled-components';
-import { useSetRecoilState, useRecoilValue } from 'recoil';
+import { useSetRecoilState } from 'recoil';
 
 import { isOpenRoomModalState } from '@atoms/is-open-modal';
-import followState from '@atoms/following-list';
 import UserCardList from '@components/common/user-card-list';
 import FollowerSelectRoomHeader from '@components/room/follower-select-room-header';
 import { BackgroundWrapper, ModalBox } from '@common/modal';
 import {
   SelectDiv, SelectInputBar, SelectedUserDiv, SelectUserComponent,
 } from '@common/select';
-import { findUsersByIdList } from '@api/index';
 import { hiddenScroll } from '@styles/scrollbar-style';
-import { IUser } from '@interfaces/index';
+import useSelectUser from '@src/hooks/useSelectUser';
 
 const Layout = styled.div`
   background-color: #F1F0E4;
@@ -45,47 +43,8 @@ const CustomInputTitle = styled.p`
 
 function FollowerSelectModal() {
   const setIsOpenRoomModal = useSetRecoilState(isOpenRoomModalState);
-  const followingList = useRecoilValue(followState);
-  const [selectedUsers, setSelectedUsers] = useState<Array<IUser>>([]);
-  const [allUserList, setAllUserList] = useState([]);
-  const [filteredUserList, setFilteredUserList] = useState([]);
-  const inputBarRef = useRef(null);
-  const selectedUserDivRef = useRef(null);
-
-  const addSelectedUser = (e: any) => {
-    const userCardDiv = e.target.closest('.userCard');
-    const profileUrl = userCardDiv.querySelector('img');
-    if (!userCardDiv || !profileUrl) return;
-    const userName = userCardDiv?.getAttribute('data-username');
-    (inputBarRef!.current as any).value = '';
-    setSelectedUsers([...selectedUsers, { userDocumentId: userCardDiv?.getAttribute('data-id'), userName, profileUrl: profileUrl.getAttribute('src') }]);
-  };
-
-  const deleteUser = (e: any) => {
-    setSelectedUsers(selectedUsers.filter((user: any) => user.userDocumentId !== e.target.getAttribute('data-id')));
-    (inputBarRef!.current as any).value = '';
-  };
-
-  const searchUser = () => {
-    const searchWord = (inputBarRef.current as any).value;
-    const selectedUserIds = selectedUsers.map((user: any) => user.userDocumentId);
-    setFilteredUserList(allUserList
-      .filter((user: any) => (user.userId.indexOf(searchWord) > -1 || user.userName.indexOf(searchWord) > -1) && selectedUserIds.indexOf(user._id) === -1));
-  };
-
-  useEffect(() => {
-    findUsersByIdList(followingList).then((res: any) => {
-      const selectedUserIds = selectedUsers.map((user: any) => user.userDocumentId);
-      setAllUserList(res.userList);
-      setFilteredUserList(res.userList.filter((user: any) => selectedUserIds.indexOf(user._id) === -1));
-    });
-  }, [followingList]);
-
-  useEffect(() => {
-    const selectedUserIds = selectedUsers.map((user: any) => user.userDocumentId);
-    setFilteredUserList(allUserList.filter((user: any) => selectedUserIds.indexOf(user._id) === -1));
-
-  }, [selectedUsers]);
+  const inputBarRef = useRef<HTMLInputElement>(null);
+  const [selectedUsers, filteredUserList, searchUser, deleteUser, addSelectedUser] = useSelectUser(inputBarRef);
 
   return (
     <>
@@ -97,7 +56,7 @@ function FollowerSelectModal() {
             <CustomInputTitle>ADD : </CustomInputTitle>
             <CustomSelectInputBar ref={inputBarRef} onChange={searchUser} />
           </SelectDiv>
-          <SelectedUserDiv ref={selectedUserDivRef}>
+          <SelectedUserDiv>
             {selectedUsers.map((user: any) => (
               <SelectUserComponent key={user.userDocumentId} data-id={user.userDocumentId} onClick={deleteUser}>
                 {user.userName}
